refactor: move all pkg to internal
We're not currently supporting a public Go API, so it doesn't make sense to use pkg at the moment.
We may open this in the future but the current integration point is using the json output using `infracost --output=json`.
