+++ internal/output/json.go
@@ -3,7 +3,7 @@ package output
 import (
 	"encoding/json"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 	"github.com/urfave/cli/v2"
 )
