+++ internal/prices/prices.go
@@ -1,7 +1,7 @@
 package prices
 
 import (
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 
 	"github.com/shopspring/decimal"
 	log "github.com/sirupsen/logrus"
