+++ internal/providers/terraform/aws/autoscaling_group.go
@@ -3,7 +3,7 @@ package aws
 import (
 	"fmt"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/shopspring/decimal"
