+++ internal/providers/terraform/aws/docdb_cluster_instance.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 	"github.com/shopspring/decimal"
 )
