+++ internal/providers/terraform/aws/docdb_cluster_instance_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/infracost/infracost/internal/providers/terraform/tftest"
-	"github.com/infracost/infracost/pkg/testutil"
+	"github.com/infracost/infracost/internal/testutil"
 	"github.com/shopspring/decimal"
 )
