+++ internal/providers/terraform/aws/elasticsearch_domain.go
@@ -3,7 +3,7 @@ package aws
 import (
 	"fmt"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 	"github.com/shopspring/decimal"
 )
