+++ internal/providers/terraform/aws/elasticsearch_domain_test.go
@@ -3,7 +3,7 @@ package aws_test
 import (
 	"testing"
 
-	"github.com/infracost/infracost/pkg/testutil"
+	"github.com/infracost/infracost/internal/testutil"
 	"github.com/shopspring/decimal"
 
 	"github.com/infracost/infracost/internal/providers/terraform/tftest"
