+++ internal/providers/terraform/aws/lambda_function.go
@@ -1,7 +1,7 @@
 package aws
 
 import (
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 
 	"github.com/shopspring/decimal"
 )
