+++ internal/providers/terraform/resource_registry.go
@@ -3,7 +3,7 @@ package terraform
 import (
 	"sync"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 
 	"github.com/infracost/infracost/internal/providers/terraform/aws"
 )
