+++ internal/schema/registry_item.go
