+++ internal/schema/resource.go
