+++ internal/testutil/testutil.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/infracost/infracost/pkg/schema"
+	"github.com/infracost/infracost/internal/schema"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/shopspring/decimal"
